Add file
add registerControoler.js and registerModel.js
